move funcs to util
