+++ x/staking/alias.go
@@ -182,7 +182,6 @@ var (
 
 	// variable aliases
 	ModuleCdc                        = types.ModuleCdc
-	PoolKey                          = types.PoolKey
 	LastValidatorPowerKey            = types.LastValidatorPowerKey
 	LastTotalPowerKey                = types.LastTotalPowerKey
 	ValidatorsKey                    = types.ValidatorsKey
